refactor(github): split 1017-line god file
Extract package-level functions into focused files:
- pr.go: PR fetch/state/files/branch/context/merge/edit ops
- review.go: FetchReviews, HasPendingReview, ApprovePR
- renovate.go: FetchRenovatePRs and GQL query
- checks.go: RerunFailedChecks
- issues.go: FetchAssignedIssues, FetchLabeledIssuesForRepos,
  FetchIssue with GQL types
- client.go: shared execer, GQL infra, convertPRs, URL parsers

Test files mirror the split: pr_test.go, review_test.go,
renovate_test.go; client_test.go retains shared helpers
and infra tests.
